ai_code_review_bot/requirements.txt
openai>=0.27
PyGithub>=2.0